lara_1.1

+++ LaravelDeneme/database/migrations/2021_03_26_135859_services_lang.php
@@ -16,7 +16,7 @@ class ServicesLang extends Migration
         Schema::create('services_lang', function (Blueprint $table) {
             $table->increments('id');
             $table->string('service_name',255);
-            $table->smallInteger('default_service_id')->unsigned();
+            $table->integer('default_service_id')->unsigned();
             $table->foreign('default_service_id')->references('id')->on('services');
             $table->string('lang',255);
             $table->string('slug',255);
